# Nic_Jacson_Course/swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
    Product:
        description: Product defines the structure for an API product
        properties:
            description:
                type: string
                x-go-name: Description
            id:
                description: The id for this user
                format: int64
                minimum: 1
                type: integer
                x-go-name: ID
            name:
                type: string
                x-go-name: Name
            price:
                format: float
                type: number
                x-go-name: Price
            sku:
                type: string
                x-go-name: SKU
        required:
            - id
        type: object
        x-go-package: github.com/sunilmaharana/data
info:
    description: '# Documentation for Product API'
    title: of Product API
    version: 1.0.0
paths:
    /products:
        get:
            description: Returns a list of products
            operationId: listProducts
            responses:
                "200":
                    $ref: '#/responses/productsResponse'
            tags:
                - products
    /products/{id}:
        delete:
            description: Deletes a product from the database
            operationId: deleteProduct
            parameters:
                - description: The id of the product to delete from the database
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "201":
                    $ref: '#/responses/noContentResponse'
            tags:
                - products
produces:
    - application/json
responses:
    noContentResponse:
        description: ""
    productsResponse:
        description: A list of products returns in response
        schema:
            items:
                $ref: '#/definitions/Product'
            type: array
schemes:
    - http
swagger: "2.0"